Move parser.rb into lib/
