+++ lib/annotate/parser.rb
